added schema and data for testing in h2 database still a work in progress

+++ src/main/java/com/morogoyodesings/CRM/repositories/CustomerRepository.java
@@ -4,6 +4,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.morogoyodesings.CRM.data.CustomerInfo;
 
-public interface CustomerRepository extends JpaRepository<CustomerInfo, Long> {
+
+
+public interface CustomerRepository extends JpaRepository<CustomerInfo, Integer> {
 
 }
